tests/wpt/web-platform-tests/html/webappapis/dynamic-markup-insertion/document-write/module-static-import-delayed.html

<!doctype html>
<title>document.write in an imported module</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(t => {
  // Expose {test} in the iframe for using the step_timeout helper.
  document.test = t;
  const iframe = document.createElement("iframe");

  iframe.onerror = t.unreached_func("Error loading iframe");
  let onLoadWasCalled = false;
  iframe.onload = t.step_func(() => {
    assert_equals(iframe.contentDocument.body.textContent, "Initial body contents\n");
    iframe.onload = null;
    onLoadWasCalled = true;
  });
  document.addEventListener("documentWriteDone", t.step_func_done(() => {
    assert_true(onLoadWasCalled);
    assert_equals(iframe.contentDocument.body.textContent, "document.write body contents\n");
  }));

  iframe.src = "module-static-import-delayed-iframe.html";
  document.body.appendChild(iframe);
});
</script>
